Refactor Neuro code to use more OOP with add tasks and bye functionality
