--- src/main/java/Storage.java
@@ -1,2 +1,0 @@
-public class Storage {
-}
