--- src/main/java/TaskList.java
@@ -1,2 +1,0 @@
-public class TaskList {
-}
