--- src/main/java/Ui.java
@@ -1,2 +1,0 @@
-public class Ui {
-}
