+++ src/main/java/neuro/Storage.java
@@ -1,0 +1,11 @@
+package neuro;
+
+public class Storage {
+    public Storage(String filePath) {
+
+    }
+
+    public String load() {
+        return "";
+    }
+}
