+++ src/main/java/neuro/Ui.java
@@ -1,0 +1,84 @@
+package neuro;
+
+import java.util.Scanner;
+
+public class Ui {
+    private final Scanner scanner;
+
+    public Ui() {
+        this.scanner = new Scanner(System.in);
+    }
+
+    public void showLine() {
+        System.out.println("    ___________________________________________________");
+    }
+    public void showWelcome() {
+        showLine();
+        System.out.println("    Hii, I'm Neuro, your chatbot assistant!");
+        System.out.println("    How can I help you?");
+        showLine();
+    }
+
+    public void showGoodbye() {
+        System.out.println("    Goodbyeee!");
+        System.out.println("""
+
+
+                                                                   @@=+==%@
+                                                                    @%==#@
+                                                                      @++@
+                                                      @@@@@@@%%%%%@@@@@@#@
+                                                 @@%*+=======+++++====++#@@
+                                             @@#+==+=++=================+==+*@@
+                                     @@@%%@@*+=+===+=#*+=====================++*@@
+                       @    @ @@%####@*#@+==+====++@+===========================++@@
+                  @%@ %-=@@*-*%#####*@#========+=@++=+====++======================+=#@
+                 @*--%*--*@---@*###%#=========+%+==+#@#++#@%========================+=*@
+                @@%---#--=*---@*##@++=+======#*=++@+=+====+===========================+=#@
+                +-*#----------##%@#%++=====+%+==+@*+========================+============+@
+                @=-----=====---==--*#====++%===**%=====++=+=============+*%@%*=+==++=====++
+                 @*---======------%+=====+#=+=%=%+====+@+=#++=============+==+@#==+#=======
+                @*#+------------=%*%*===*#==#*=+#====#*@==@++=+================#++=@=======
+                ==#@#+======+#%%#===+#+##=##=--%=++=@=-@++@++====++====++========+=@++=====
+                #=======--===========%+###=---#++=%+=--#++@+========+==%@========+=@*+=====
+                %-===================*%=-----=#=#*=----+*+#*=======+@=#+@=========+%*+=====
+                %-===================++------%#%=-------#**#=++===+@#**=%=========*##+=====
+                @=====================#------*=---------#**%=====+##+#-=#=========%##======
+                @+====================@@@@@@*-=#+-------+%##*+=++@=%%=-+*=========@#*======
+                @#====================@@@@@@@@%=-%-------%%+#===%=+@=--*+========+*#+======
+                @%====================%*%@@@@@@@%=-------+@=#=+#*-*%---%=+=======#-%=======
+                @@====================#+====*@@@@@*-------@=+#=%=-#*--=%=======++#=#=======
+                %*+===================**======+@@#-----------%+#------=#@+=++=+=%=#++======
+                +=#====================#========%=-----------=@+---==+*#@+=#%++@==@========
+                +=%===================-#========-----------------#@@@@@@@%*=--===*+========
+                =+%+==================-#========--=+=----------+@@@@@@@@@@@@@@*-=%=========
+                =+**===================#+=====----%**#%%*=-----+*+===+*%@@@@@@@@%++========
+                ===#-==================**=-------##*******%#------========*@@@@@*+=========
+                =+*@====================#--------%**********-----============*@*=========+*
+                =+@@*==================-%--------%********@=----=============%+=========+=%
+                +@ @%===================*--------**======#=-----===========*%============%*
+                @   @===================+@=-------+====+%=-------=======+%#*@=+========++%*
+                    @*===================@-+@*=-------------------=========@++========++@**
+                   @%@===================#==+*-+%#+=-----------------=====%=+==========%***
+                   @*%+===================*==#*==+*=+###*==----------=*#%%============##***
+                  @#**#-=================-%===#*=-%#%*=*#===++++#@#****#*=+==========*#****
+                 @#***%+==================+*=*++*-*==*#====-=##+*#****%+=+%%+=======*%*****
+                @%*****%-=================-#-%+*%%+#%=-=#%#+=====+%*%@@%**%=======+##******
+                %******##==================+%@++=%@@@#%%%*=========%*****%======+=@********
+
+                """);
+    }
+
+    public void showMessage(String message) {
+        System.out.println("    " + message);
+    }
+
+    public void showError(String errorMessage) {
+        System.out.println("    Oops! " + errorMessage);
+    }
+
+    public String readCommand() {
+        System.out.print("$ ");
+        return scanner.nextLine();
+    }
+}
