+++ src/main/java/neuro/command/DeleteCommand.java
@@ -1,0 +1,17 @@
+package neuro.command;
+import neuro.Ui;
+import neuro.Storage;
+
+import neuro.task.TaskList;
+
+public class DeleteCommand extends Command {
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+
+    }
+
+    @Override
+    public boolean isExit() {
+        return false;
+    }
+}
