+++ src/main/java/neuro/command/ListCommand.java
@@ -1,0 +1,4 @@
+package neuro.command;
+
+public class ListCommand {
+}
