+++ src/main/java/neuro/command/MarkCommand.java
@@ -1,0 +1,4 @@
+package neuro.command;
+
+public class MarkCommand {
+}
