+++ src/main/java/neuro/command/UnmarkCommand.java
@@ -1,0 +1,4 @@
+package neuro.command;
+
+public class UnmarkCommand {
+}
