+++ src/main/java/neuro/task/Deadline.java
@@ -1,3 +1,7 @@
+package neuro.task;
+
+import neuro.task.Task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
