+++ src/main/java/neuro/task/Event.java
@@ -1,3 +1,7 @@
+package neuro.task;
+
+import neuro.task.Task;
+
 public class Event extends Task {
     private String from;
     private String to;
