+++ src/main/java/neuro/task/Task.java
@@ -1,3 +1,5 @@
+package neuro.task;
+
 public abstract class Task {
     private final String description;
     private boolean isDone;
