+++ src/main/java/neuro/task/Todo.java
@@ -1,3 +1,7 @@
+package neuro.task;
+
+import neuro.task.Task;
+
 public class Todo extends Task {
     public Todo(String description) {
         super(description);
