+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
     ___________________________________________________
-    Hii, I'm Neuro, your chatbot assistant!
+    Hii, I'm neuro.Neuro, your chatbot assistant!
     How can I help you?
     ___________________________________________________
     ___________________________________________________
